:wastebasket: update delete_intervention

+++ delete_inter.php
@@ -3,13 +3,14 @@
 header("Content-Type: application/json; charset=UTF-8");
 
 require_once 'connect.php';
-
-$request_info = $_GET['request_info'];
-$request_id      = $_GET['request_id'];
-
-
+$key = $_POST['key'];
+$request_info = $_POST['request_info'];
+$request_id      = $_POST['request_id'];
 
 
+
+
+if ( $key == "delete" ){
     $query = "DELETE FROM assignwork_tb WHERE request_id='$request_id' ";
 
         if ( mysqli_query($conn, $query) ){
@@ -33,6 +34,13 @@ $request_id      = $_GET['request_id'];
             mysqli_close($conn);
         }
 
+    } else {
+        $response["value"] = "0";
+        $response["message"] = "Error! ".mysqli_error($conn);
+        echo json_encode($response);
+    
+        mysqli_close($conn);
+    }
 
 
 ?>
